+++ jsonschema-generator/src/main/java/com/github/victools/jsonschema/generator/impl/SchemaGenerationContextImpl.java
@@ -293,7 +293,7 @@ public class SchemaGenerationContextImpl implements SchemaGenerationContext {
                 includeTypeAttributes = customDefinition.shouldIncludeAttributes();
             } else if (isContainerType) {
                 logger.debug("generating array definition for {}", targetType);
-                this.generateArrayDefinition(targetType, definition, isNullable);
+                this.generateArrayDefinition(scope, definition, isNullable);
                 includeTypeAttributes = true;
             } else {
                 logger.debug("generating definition for {}", targetType);
@@ -339,10 +339,10 @@ public class SchemaGenerationContextImpl implements SchemaGenerationContext {
     }
 
     /**
-     * Collect the specified value(s) from the given definition's "{@value SchemaConstants#TAG_TYPE}" attribute.
+     * Collect the specified value(s) from the given definition's {@link SchemaKeyword#TAG_TYPE} attribute.
      *
-     * @param definition type definition to extract specified "{@value SchemaConstants#TAG_TYPE}" values from
-     * @return extracted "{@value SchemaConstants#TAG_TYPE}" values (may be empty)
+     * @param definition type definition to extract specified {@link SchemaKeyword#TAG_TYPE} values from
+     * @return extracted {@link SchemaKeyword#TAG_TYPE} – values (may be empty)
      */
     private Set<String> collectAllowedSchemaTypes(ObjectNode definition) {
         JsonNode declaredTypes = definition.get(this.getKeyword(SchemaKeyword.TAG_TYPE));
@@ -362,11 +362,11 @@ public class SchemaGenerationContextImpl implements SchemaGenerationContext {
     /**
      * Preparation Step: add the given targetType (which was previously determined to be an array type).
      *
-     * @param targetType (possibly generic) array type to add
+     * @param targetScope (possibly generic) array type to add
      * @param definition node in the JSON schema to which all collected attributes should be added
      * @param isNullable whether the field/method's return value the targetType refers to is allowed to be null in the declaring type
      */
-    private void generateArrayDefinition(ResolvedType targetType, ObjectNode definition, boolean isNullable) {
+    private void generateArrayDefinition(TypeScope targetScope, ObjectNode definition, boolean isNullable) {
         if (isNullable) {
             ArrayNode typeArray = this.generatorConfig.createArrayNode()
                     .add(this.getKeyword(SchemaKeyword.TAG_TYPE_ARRAY))
@@ -375,10 +375,22 @@ public class SchemaGenerationContextImpl implements SchemaGenerationContext {
         } else {
             definition.put(this.getKeyword(SchemaKeyword.TAG_TYPE), this.getKeyword(SchemaKeyword.TAG_TYPE_ARRAY));
         }
-        ObjectNode arrayItemTypeRef = this.generatorConfig.createObjectNode();
-        definition.set(this.getKeyword(SchemaKeyword.TAG_ITEMS), arrayItemTypeRef);
-        ResolvedType itemType = this.typeContext.getContainerItemType(targetType);
-        this.traverseGenericType(itemType, arrayItemTypeRef, false);
+        if (targetScope instanceof MemberScope<?, ?> && !((MemberScope<?, ?>) targetScope).isFakeContainerItemScope()) {
+            MemberScope<?, ?> fakeArrayItemMember = ((MemberScope<?, ?>) targetScope).asFakeContainerItemScope();
+            Map<String, JsonNode> fakeItemDefinition = new HashMap<>();
+            if (targetScope instanceof FieldScope) {
+                this.populateField((FieldScope) fakeArrayItemMember, fakeItemDefinition, null);
+            } else if (targetScope instanceof MethodScope) {
+                this.populateMethod((MethodScope) fakeArrayItemMember, fakeItemDefinition, null);
+            } else {
+                throw new IllegalStateException("Unsupported member type: " + targetScope.getClass().getName());
+            }
+            definition.set(this.getKeyword(SchemaKeyword.TAG_ITEMS), fakeItemDefinition.values().iterator().next());
+        } else {
+            ObjectNode arrayItemTypeRef = this.generatorConfig.createObjectNode();
+            definition.set(this.getKeyword(SchemaKeyword.TAG_ITEMS), arrayItemTypeRef);
+            this.traverseGenericType(targetScope.getContainerItemType(), arrayItemTypeRef, false);
+        }
     }
 
     /**
@@ -495,18 +507,28 @@ public class SchemaGenerationContextImpl implements SchemaGenerationContext {
      * @param requiredProperties set of properties value required
      */
     private void populateField(FieldScope field, Map<String, JsonNode> collectedFields, Set<String> requiredProperties) {
-        String propertyNameOverride = this.generatorConfig.resolvePropertyNameOverride(field);
-        FieldScope fieldWithNameOverride = propertyNameOverride == null ? field : field.withOverriddenName(propertyNameOverride);
-        String propertyName = fieldWithNameOverride.getSchemaPropertyName();
-        if (this.generatorConfig.isRequired(field)) {
-            requiredProperties.add(propertyName);
-        }
-        if (collectedFields.containsKey(propertyName)) {
-            logger.debug("ignoring overridden {}.{}", fieldWithNameOverride.getDeclaringType(), fieldWithNameOverride.getDeclaredName());
-            return;
+        final FieldScope fieldWithNameOverride;
+        final String propertyName;
+        if (field.isFakeContainerItemScope()) {
+            fieldWithNameOverride = field;
+            propertyName = field.getSchemaPropertyName();
+        } else {
+            String propertyNameOverride = this.generatorConfig.resolvePropertyNameOverride(field);
+            fieldWithNameOverride = propertyNameOverride == null ? field : field.withOverriddenName(propertyNameOverride);
+            propertyName = fieldWithNameOverride.getSchemaPropertyName();
+            if (this.generatorConfig.isRequired(field)) {
+                requiredProperties.add(propertyName);
+            }
+            if (collectedFields.containsKey(propertyName)) {
+                logger.debug("ignoring overridden {}.{}", fieldWithNameOverride.getDeclaringType(), fieldWithNameOverride.getDeclaredName());
+                return;
+            }
         }
 
         List<ResolvedType> typeOverrides = this.generatorConfig.resolveTargetTypeOverrides(fieldWithNameOverride);
+        if (typeOverrides == null) {
+            typeOverrides = this.generatorConfig.resolveSubtypes(fieldWithNameOverride.getType(), this);
+        }
         List<FieldScope> fieldOptions;
         if (typeOverrides == null || typeOverrides.isEmpty()) {
             fieldOptions = Collections.singletonList(fieldWithNameOverride);
@@ -516,7 +538,7 @@ public class SchemaGenerationContextImpl implements SchemaGenerationContext {
                     .collect(Collectors.toList());
         }
         // consider declared type (instead of overridden one) for determining null-ability
-        boolean isNullable = !field.getRawMember().isEnumConstant() && this.generatorConfig.isNullable(field);
+        boolean isNullable = !field.getRawMember().isEnumConstant() && !field.isFakeContainerItemScope() && this.generatorConfig.isNullable(field);
         if (fieldOptions.size() == 1) {
             collectedFields.put(propertyName, this.createFieldSchema(fieldOptions.get(0), isNullable, null));
         } else {
@@ -524,8 +546,8 @@ public class SchemaGenerationContextImpl implements SchemaGenerationContext {
             collectedFields.put(propertyName, subSchema);
             ArrayNode anyOfArray = subSchema.withArray(this.getKeyword(SchemaKeyword.TAG_ANYOF));
             if (isNullable) {
-                anyOfArray.add(this.generatorConfig.createObjectNode()
-                        .put(this.getKeyword(SchemaKeyword.TAG_TYPE), this.getKeyword(SchemaKeyword.TAG_TYPE_NULL)));
+                anyOfArray.addObject()
+                        .put(this.getKeyword(SchemaKeyword.TAG_TYPE), this.getKeyword(SchemaKeyword.TAG_TYPE_NULL));
             }
             fieldOptions.forEach(option -> anyOfArray.add(this.createFieldSchema(option, false, null)));
         }
@@ -555,18 +577,28 @@ public class SchemaGenerationContextImpl implements SchemaGenerationContext {
      * @param requiredProperties set of properties value required
      */
     private void populateMethod(MethodScope method, Map<String, JsonNode> collectedMethods, Set<String> requiredProperties) {
-        String propertyNameOverride = this.generatorConfig.resolvePropertyNameOverride(method);
-        MethodScope methodWithNameOverride = propertyNameOverride == null ? method : method.withOverriddenName(propertyNameOverride);
-        String propertyName = methodWithNameOverride.getSchemaPropertyName();
-        if (this.generatorConfig.isRequired(method)) {
-            requiredProperties.add(propertyName);
-        }
-        if (collectedMethods.containsKey(propertyName)) {
-            logger.debug("ignoring overridden {}.{}", methodWithNameOverride.getDeclaringType(), methodWithNameOverride.getDeclaredName());
-            return;
+        final MethodScope methodWithNameOverride;
+        final String propertyName;
+        if (method.isFakeContainerItemScope()) {
+            methodWithNameOverride = method;
+            propertyName = method.getSchemaPropertyName();
+        } else {
+            String propertyNameOverride = this.generatorConfig.resolvePropertyNameOverride(method);
+            methodWithNameOverride = propertyNameOverride == null ? method : method.withOverriddenName(propertyNameOverride);
+            propertyName = methodWithNameOverride.getSchemaPropertyName();
+            if (this.generatorConfig.isRequired(method)) {
+                requiredProperties.add(propertyName);
+            }
+            if (collectedMethods.containsKey(propertyName)) {
+                logger.debug("ignoring overridden {}.{}", methodWithNameOverride.getDeclaringType(), methodWithNameOverride.getDeclaredName());
+                return;
+            }
         }
 
         List<ResolvedType> typeOverrides = this.generatorConfig.resolveTargetTypeOverrides(methodWithNameOverride);
+        if (typeOverrides == null && !methodWithNameOverride.isVoid()) {
+            typeOverrides = this.generatorConfig.resolveSubtypes(methodWithNameOverride.getType(), this);
+        }
         List<MethodScope> methodOptions;
         if (typeOverrides == null || typeOverrides.isEmpty()) {
             methodOptions = Collections.singletonList(methodWithNameOverride);
@@ -576,7 +608,8 @@ public class SchemaGenerationContextImpl implements SchemaGenerationContext {
                     .collect(Collectors.toList());
         }
         // consider declared type (instead of overridden one) for determining null-ability
-        boolean isNullable = methodWithNameOverride.isVoid() || this.generatorConfig.isNullable(methodWithNameOverride);
+        boolean isNullable = methodWithNameOverride.isVoid()
+                || !method.isFakeContainerItemScope() && this.generatorConfig.isNullable(methodWithNameOverride);
         if (methodOptions.size() == 1) {
             collectedMethods.put(propertyName, this.createMethodSchema(methodOptions.get(0), isNullable, null));
         } else {
@@ -627,8 +660,16 @@ public class SchemaGenerationContextImpl implements SchemaGenerationContext {
         final CustomDefinition customDefinition = this.generatorConfig.getCustomDefinition(scope, this, ignoredDefinitionProvider);
         if (customDefinition != null && customDefinition.isMeantToBeInline()) {
             targetNode.setAll(customDefinition.getValue());
-            if (customDefinition.shouldIncludeAttributes() && collectedAttributes != null && collectedAttributes.size() > 0) {
-                targetNode.setAll(collectedAttributes);
+            if (customDefinition.shouldIncludeAttributes()) {
+                if (collectedAttributes != null && collectedAttributes.size() > 0) {
+                    targetNode.setAll(collectedAttributes);
+                }
+                Set<String> allowedSchemaTypes = this.collectAllowedSchemaTypes(targetNode);
+                ObjectNode typeAttributes = AttributeCollector.collectTypeAttributes(scope, this, allowedSchemaTypes);
+                // ensure no existing attributes in the 'definition' are replaced, by way of first overriding any conflicts the other way around
+                typeAttributes.setAll(targetNode);
+                // apply merged attributes
+                targetNode.setAll(typeAttributes);
             }
             if (isNullable) {
                 this.makeNullable(targetNode);
@@ -670,14 +711,14 @@ public class SchemaGenerationContextImpl implements SchemaGenerationContext {
             // cannot be sure what is specified in those other schema parts, instead simply create a oneOf wrapper
             ObjectNode nullSchema = this.generatorConfig.createObjectNode()
                     .put(this.getKeyword(SchemaKeyword.TAG_TYPE), this.getKeyword(SchemaKeyword.TAG_TYPE_NULL));
-            ArrayNode oneOf = this.generatorConfig.createArrayNode()
+            ArrayNode anyOf = this.generatorConfig.createArrayNode()
                     // one option in the oneOf should be null
                     .add(nullSchema)
                     // the other option is the given (assumed to be) not-nullable node
                     .add(this.generatorConfig.createObjectNode().setAll(node));
             // replace all existing (and already copied properties with the oneOf wrapper
             node.removeAll();
-            node.set(this.getKeyword(SchemaKeyword.TAG_ONEOF), oneOf);
+            node.set(this.getKeyword(SchemaKeyword.TAG_ANYOF), anyOf);
         } else {
             // given node is a simple schema, we can simply adjust its "type" attribute
             JsonNode fixedJsonSchemaType = node.get(this.getKeyword(SchemaKeyword.TAG_TYPE));