+++ jsonschema-module-swagger-1.5/src/main/java/com/github/victools/jsonschema/module/swagger15/SwaggerModule.java
@@ -123,6 +123,9 @@ public class SwaggerModule implements Module {
      * @return description (or {@code null})
      */
     protected String resolveDescription(MemberScope<?, ?> member) {
+        if (member.isFakeContainerItemScope()) {
+            return null;
+        }
         return Optional.ofNullable(member.getAnnotationConsideringFieldAndGetter(ApiModelProperty.class))
                 .map(ApiModelProperty::value)
                 .filter(value -> !value.isEmpty())
@@ -164,6 +167,9 @@ public class SwaggerModule implements Module {
      * @return {@link ApiModelProperty} annotation's non-empty {@code allowableValues} (or {@code null})
      */
     private Optional<String> findModelPropertyAllowableValues(MemberScope<?, ?> member) {
+        if (member.isFakeContainerItemScope()) {
+            return null;
+        }
         return Optional.ofNullable(member.getAnnotationConsideringFieldAndGetter(ApiModelProperty.class))
                 .map(ApiModelProperty::allowableValues)
                 .filter(allowableValues -> !allowableValues.isEmpty());